fix soundness issues in unsafe usage
This patch fixes some soundness issues in the unsafe usage.
These issues were unlikely to be able to cause issues in
practice, but could potentially cause issues if doing
something weird like mixing and matching toolchains.

It manages to significantly cut down on the unsafe usage
required, so it should make auditing the crate easier in
the future as well.